+++ src/ExponentialFamily.jl
@@ -37,7 +37,7 @@ include("distributions/lognormal.jl")
 include("distributions/binomial.jl")
 # include("distributions/multinomial.jl")
 include("distributions/wishart.jl")
-# include("distributions/wishart_inverse.jl")
+include("distributions/wishart_inverse.jl")
 # include("distributions/contingency.jl")
 include("distributions/erlang.jl")
 include("distributions/exponential.jl")
